pyproject.toml
[project]
name = "chatbot-demo"
version = "0.1.0"
description = "RAG 기반 대화형 AI Chatbot MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "langchain",
    "langchain-chroma",
    "langchain-community",
    "langchain-openai",
    "langchain-text-splitters",
    "chromadb",
    "python-dotenv",
    "streamlit",
    "sse-starlette",
    "httpx",
    "langsmith",
    "langgraph",
    "langgraph-cli[inmem]>=0.4.11",
]
